Add permissions to the REST methods based on the voters.

+++ src/Api/Controller/ReservationController.php
@@ -8,6 +8,7 @@ use App\Entity\Reservation;
 use App\Service\AppUserManager;
 use App\Service\ReservationManager;
 use App\Service\RoomManager;
+use App\Voter\ReservationVoter;
 use Exception;
 use FOS\RestBundle\Controller\AbstractFOSRestController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
@@ -29,6 +30,7 @@ class ReservationController extends AbstractFOSRestController
     #[Rest\View(statusCode: 200)]
     public function list(Request $request): array
     {
+        $this->denyAccessUnlessGranted(ReservationVoter::VIEW_INDEX);
         $title = $request->query->get('title');
         $description = $request->query->get('description');
 
@@ -50,6 +52,7 @@ class ReservationController extends AbstractFOSRestController
     public function detail(int $id): ReservationOutput
     {
         $reservation = $this->reservationManager->findById($id);
+        $this->denyAccessUnlessGranted(ReservationVoter::VIEW_DETAIL, $reservation);
         if (!$reservation) {
             throw new HttpException(404, 'Reservation not found');
         }
@@ -71,7 +74,13 @@ class ReservationController extends AbstractFOSRestController
     #[Rest\View(statusCode: 201)]
     public function update(?int $id, ReservationInput $reservationInput, ConstraintViolationListInterface $errors): ReservationOutput
     {
-        $reservation = $id !== null ? $this->findOrFail($id) : new Reservation();
+        if ($id !== null) {
+            $reservation = $this->findOrFail($id);
+            $this->denyAccessUnlessGranted(ReservationVoter::EDIT, $reservation);
+        } else {
+            $reservation = new Reservation();
+            $this->denyAccessUnlessGranted(ReservationVoter::CREATE);
+        }
 
         if ($errors->count() > 0) {
             throw new HttpException(400, message: \implode("\n", \array_map(
@@ -95,6 +104,7 @@ class ReservationController extends AbstractFOSRestController
     #[Rest\View(statusCode: 204)]
     public function delete(int $id): void
     {
+        $this->denyAccessUnlessGranted(ReservationVoter::DELETE);
         $reservation = $this->findOrFail($id);
         $this->reservationManager->deleteFromDatabase($reservation);
     }
@@ -107,6 +117,7 @@ class ReservationController extends AbstractFOSRestController
     public function approve(int $id): ReservationOutput
     {
         $reservation = $this->findOrFail($id);
+        $this->denyAccessUnlessGranted(ReservationVoter::CAN_APPROVE, $reservation);
 
         $reservation->setStatus('approved');
         $reservation = $this->reservationManager->saveToDatabase($reservation);
@@ -127,6 +138,7 @@ class ReservationController extends AbstractFOSRestController
     public function reject(int $id): ReservationOutput
     {
         $reservation = $this->findOrFail($id);
+        $this->denyAccessUnlessGranted(ReservationVoter::CAN_REJECT, $reservation);
 
         $reservation->setStatus('rejected');
         $reservation = $this->reservationManager->saveToDatabase($reservation);
